+++ src/applications/disability-benefits/996/components/SelectArrayItemsWidget.jsx
